+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -2,7 +2,7 @@ package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
+import webmvc.org.springframework.web.servlet.mvc.handler.asis.Controller;
 
 public class RegisterViewController implements Controller {
