+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/frontcontroller/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.disapatchersevlet;
+package webmvc.org.springframework.web.servlet.mvc.frontcontroller;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
