+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/frontcontroller/HandlerMappings.java
@@ -1,9 +1,9 @@
-package webmvc.org.springframework.web.servlet.mvc.disapatchersevlet;
+package webmvc.org.springframework.web.servlet.mvc.frontcontroller;
 
 import jakarta.servlet.http.HttpServletRequest;
-import webmvc.org.springframework.web.servlet.mvc.asis.ControllerHandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.asis.ForwardController;
-import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.handler.tobe.controllerhandler.ControllerHandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.handler.asis.ForwardController;
+import webmvc.org.springframework.web.servlet.mvc.handler.tobe.annoationhandler.AnnotationHandlerMapping;
 
 import java.util.ArrayList;
 import java.util.List;
